RacingGameApplication 시작 후 입력받는 부분을 분리
